+++ src/main/java/act/cli/CliSession.java
@@ -31,6 +31,8 @@ import act.util.Banner;
 import act.util.DestroyableBase;
 import jline.Terminal;
 import jline.console.ConsoleReader;
+import jline.console.history.FileHistory;
+import jline.console.history.PersistentHistory;
 import org.osgl.$;
 import org.osgl.logging.LogManager;
 import org.osgl.logging.Logger;
@@ -56,6 +58,7 @@ public class CliSession extends DestroyableBase implements Runnable {
     private boolean exit;
     private Thread runningThread;
     private ConsoleReader console;
+    private FileHistory history;
     private CliCursor cursor;
     private CommandNameCompleter commandNameCompleter;
     // the current handler
@@ -170,6 +173,14 @@ public class CliSession extends DestroyableBase implements Runnable {
 //        w.flush();
     }
 
+    private File historyFile() {
+        String fileName = System.getProperty("cli.history");
+        if (S.blank(fileName)) {
+            fileName = ".act.cli-history";
+        }
+        return new File(fileName);
+    }
+
     @Override
     public void run() {
         runningThread = Thread.currentThread();
@@ -178,6 +189,8 @@ public class CliSession extends DestroyableBase implements Runnable {
             OutputStream os = socket.getOutputStream();
             tryTuneTelnetOptions(os);
             console = new ConsoleReader(socket.getInputStream(), os);
+            history = new FileHistory(historyFile(), true);
+            console.setHistory(history);
             String banner = Banner.cachedBanner();
             printBanner(banner, console);
             String appName = App.instance().name();
@@ -235,6 +248,7 @@ public class CliSession extends DestroyableBase implements Runnable {
         } catch (Exception e) {
             LOGGER.error(e, "Error processing cli session");
         } finally {
+            IO.flush(history);
             if (null != server) {
                 server.remove(this);
             }